+++ projects/contexr/src/lib/components/context-menu/context-menu.component.ts
@@ -4,6 +4,7 @@ import {ContexrService} from '../../providers/contexr.service';
 import {Subscription} from 'rxjs';
 import {ContextMenuItem} from '../../types/context-menu-item';
 import {ContextMenuEntry} from '../../types/context-menu-entry';
+import {Submenu} from 'contexr/lib/types/submenu';
 
 @Component({
   selector: 'ctx-context-menu',
@@ -18,6 +19,10 @@ export class ContextMenuComponent implements OnInit, OnDestroy {
 
   constructor(public contexr: ContexrService) {
     // Event capturing (not possible in real Angular yet)
+    document.addEventListener('click', (event) => {
+      event.preventDefault();
+      this.contexr.reset();
+    }, true);
     document.addEventListener('contextmenu', (event) => {
       event.preventDefault();
       this.contexr.reset();
@@ -80,6 +85,14 @@ export class ContextMenuComponent implements OnInit, OnDestroy {
    * @returns
    */
   public isAction(item: ContextMenuEntry): boolean {
-    return !!(item as ContextMenuItem).action;
+    return !(item as ContextMenuItem).hideMenu;
+  }
+
+  /**
+   * Is this context menu entry a submenu?
+   * @param item
+   */
+  public isSubmenu(item: ContextMenuEntry): boolean {
+    return !!(item as Submenu).children;
   }
 }